pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "v6-session-basics"
version = "1.0.0"
description = "Basic data extraction and processing jobs"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "pandas",
  "vantage6-algorithm-tools==5.0.0a36",
]
authors = [
  { name = "Bart van Beusekom", email = "b.vanbeusekom@iknl.nl" },
  { name = "Frank Martin", email = "f.martin@iknl.nl" },
]
license = { text = "MIT" }
keywords = ["vantage6", "algorithm", "federated-learning"]
classifiers = [
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.13",
]

[project.urls]
Homepage = "https://github.com/vantage6/v6-session-basics"
Repository = "https://github.com/vantage6/v6-session-basics"
Issues = "https://github.com/vantage6/v6-session-basics/issues"

[tool.hatch.build.targets.wheel]
packages = ["v6-session-basics"]

[tool.setuptools.package-data]
"v6-session-basics" = ["data/*.csv"]

[tool.uv]
dev-dependencies = []
